requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
tldextract
python-dotenv

langgraph
langchain
langchain-openai
langchain-community
chromadb
langgraph
pydantic
streamlit
streamlit-chat
honcho